+++ project/settings/urls.py
@@ -17,7 +17,7 @@ Including another URLconf
 
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from common.views import HealthViewSet
+from core.views import HealthViewSet
 from users.views import AccountViewSet, AuthViewSet, SecurityQuestionViewSet
 
 router = DefaultRouter()
