+++ project/tests/users/views/test_auth_view.py
@@ -6,7 +6,7 @@ from rest_framework import status
 from rest_framework.test import APIClient
 
 from users.models import UserQuestionAnswer, CustomUser
-from common.constants import ErrorTypes, TokenScope, BYPASS_TOKEN_NOTFOUND, EventType
+from core.constants import ErrorTypes, TokenScope, BYPASS_TOKEN_NOTFOUND, EventType
 from users.redis.tokens import RefreshTokenRedis, ScopeTokenRedis
 from tests.helpers.setup_mock_accounts import (
     setup_mock_accounts,
